[PHI] move diag_embed op to phi. (#44408)
* move diag_embed to phi.
