+++ paddle/phi/api/yaml/legacy_api.yaml
@@ -524,6 +524,14 @@
     func : determinant
   backward : det_grad
 
+- api : diag_embed
+  args : (Tensor x, int offset, int dim1, int dim2)
+  output : Tensor
+  infer_meta :
+    func : DiagEmbedInferMeta
+  kernel :
+    func : diag_embed
+
 - api : divide
   args : (Tensor x, Tensor y)
   output : Tensor
