+++ python/paddle/nn/functional/extension.py
@@ -98,11 +98,17 @@ def diag_embed(input, offset=0, dim1=-2, dim2=-1):
             #  [[ 0.        ,  0.        ,  0.        ,  0.        ],
             #   [ 0.        ,  0.        ,  0.        ,  0.        ]]]
     """
-    inputs = {'Input': [input]}
-    attrs = {'offset': offset, 'dim1': dim1, 'dim2': dim2}
-
     if not isinstance(input, Variable):
         input = assign(input)
+
+    if in_dygraph_mode():
+        return _C_ops.final_state_diag_embed(input, offset, dim1, dim2)
+    elif in_dynamic_mode():
+        return _C_ops.diag_embed(input, "offset", offset, "dim1", dim1, "dim2",
+                                 dim2)
+
+    inputs = {'Input': [input]}
+    attrs = {'offset': offset, 'dim1': dim1, 'dim2': dim2}
 
     def __check_input(input, offset, dim1, dim2):
         check_dtype(input.dtype, 'Input',
@@ -129,8 +135,7 @@ def diag_embed(input, offset=0, dim1=-2, dim2=-1):
                "dim1 and dim2 cannot be the same dimension." \
                 "But received dim1 = %d, dim2 = %d\n"%(dim1, dim2)
 
-    if not in_dynamic_mode():
-        __check_input(input, offset, dim1, dim2)
+    __check_input(input, offset, dim1, dim2)
     helper = LayerHelper("diag_embed", **locals())
 
     out = helper.create_variable_for_type_inference(dtype=input.dtype)
